working-requirements.txt
anyascii==0.3.3
asgiref==3.9.1
beautifulsoup4==4.12.3
black==24.1.1
boto3==1.34.14
botocore==1.34.162
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.2.6
django-environ==0.11.2
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-storages==1.14.2
django-stubs-ext==5.2.5
django-taggit==5.0.1
django-tasks==0.6.1
django-treebeard==4.7
djangorestframework==3.16.1
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
flake8==7.0.0
gunicorn==21.2.0
html5lib==1.1
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
laces==0.1.2
mccabe==0.7.0
mypy_extensions==1.1.0
openpyxl==3.1.2
packaging==25.0
pathspec==0.12.1
pillow==10.2.0
pillow_heif==0.22.0
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.10.4
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.15.0
urllib3==2.5.0
wagtail==6.4.1
webencodings==0.5.1
whitenoise==6.6.0
Willow==1.8.0
